requirements.txt
arabic-reshaper==2.0.15
asgiref==3.2.7
autopep8==1.5
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.3
dj-database-url==0.5.0
Django==3.0.4
django-crispy-forms==1.9.0
django-heroku==0.3.1
entrypoints==0.3
flake8==3.7.9
future==0.18.2
gunicorn==20.0.4
idna==2.9
mccabe==0.6.1
Pillow==7.0.0
psycopg2==2.8.4
pycodestyle==2.5.0
pyflakes==2.1.1
python-bidi==0.4.2
pytz==2019.3
qrcode==6.1
requests==2.23.0
six==1.14.0
sqlparse==0.3.1
urllib3==1.25.8
whitenoise==5.0.1
